requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Multi-Agent System dependencies
PyJWT==2.8.0
cryptography==41.0.8

# AI Model Integrations
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# YouTube API and OAuth
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.25.2

# Database
# sqlite3 is built into Python and doesn't need to be in requirements.txt
asyncpg==0.29.0  # For enhanced database connections
redis==5.0.1  # For session management
supabase==2.3.0  # For Supabase integration

# Caching and Performance
cachetools==5.3.2

# HTTP and Networking
httpx==0.25.2
requests==2.31.0

# Date and Time
pytz==2023.3  # For timezone handling

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For advanced features
# numpy==1.24.3  # For advanced analytics
# pandas==2.1.3  # For data processing
# scikit-learn==1.3.2  # For ML features